Replace bundler used, from Webpack to Parcel
This change was necessary due to friction in running dev builds without
live reloading.

I chose Parcel instead of Rollup, the obvious choice for working with
Svelte, due to incompatibilities with near-api-js. This change also
simplified a lot of the configuration.

Some difficulties converting to Parcel stemmed from needing
an older version of PostCSS and a compatible version of tailwindCSS to
play nice with Parcel v1.